requirements.txt
django==4.2.30
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-environ==0.13.0
django-cors-headers==4.9.0
django-redis==6.0.0
psycopg2-binary==2.9.11
redis==7.0.1
